Moved module recommendation into tournesol app
