+++ backend/tournesol/recommendation/recomended_user.py
@@ -1,5 +1,5 @@
 from core.models.user import User as UserDB
-from recommendation.video import Video
+from tournesol.recommendation.video import Video
 from tournesol.models import ContributorRatingCriteriaScore, Entity, Poll
